requirements.txt
mysql-connector-python
fastapi
uvicorn
pydantic
pandas
beautifulsoup4
transformers
tensorflow
tf-keras
ray  
ray[serve]